Refactor category components and structure; introduce CategoryProfileCard, CategoryInfoCard, and CategoryItemsCard for improved organization and reusability
