+++ client/admin-web/src/pages/category/components/categories-page/CategoriesTableSkeleton.jsx
